Move JSON type annotations and JSON Schemas
